# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatModel {
  /** Internal model identifier */
  id: string;
  /** Pretty display name */
  label: string;
  /** One-liner describing the model's vibe */
  tagline: string;
  premium: boolean;
  emoji: string;
}

export interface ImageModel {
  id: string;
  label: string;
  tagline: string;
  premium: boolean;
  emoji: string;
}

export interface Persona {
  id: string;
  label: string;
  description: string;
  emoji: string;
}

export interface ModelPresets {
  chatModels: ChatModel[];
  imageModels: ImageModel[];
  personas: Persona[];
  version: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  /** Chat model id from /models */
  modelId: string;
  /** Persona id from /models */
  personaId: string;
  /** Full chat history (last 50 messages used) */
  messages: ChatMessage[];
}

export type ImageRequestSize =
  (typeof ImageRequestSize)[keyof typeof ImageRequestSize];

export const ImageRequestSize = {
  "1024x1024": "1024x1024",
  "1536x1024": "1536x1024",
  "1024x1536": "1024x1536",
} as const;

export interface ImageRequest {
  modelId: string;
  prompt: string;
  size?: ImageRequestSize;
}

export interface GeneratedImage {
  /** Base64-encoded PNG image */
  b64Json: string;
  prompt: string;
  modelId: string;
}