# lib/api-zod/src/generated/types/chatMessageRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;