# lib/api-zod/src/generated/types/chatModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatModel {
  /** Internal model identifier */
  id: string;
  /** Pretty display name */
  label: string;
  /** One-liner describing the model's vibe */
  tagline: string;
  premium: boolean;
  emoji: string;
}