# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatRequest {
  /** Chat model id from /models */
  modelId: string;
  /** Persona id from /models */
  personaId: string;
  /** Full chat history (last 50 messages used) */
  messages: ChatMessage[];
}