# lib/api-zod/src/generated/types/generatedImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratedImage {
  /** Base64-encoded PNG image */
  b64Json: string;
  prompt: string;
  modelId: string;
}