# lib/api-zod/src/generated/types/imageModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageModel {
  id: string;
  label: string;
  tagline: string;
  premium: boolean;
  emoji: string;
}