# lib/api-zod/src/generated/types/imageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageRequestSize } from "./imageRequestSize";

export interface ImageRequest {
  modelId: string;
  prompt: string;
  size?: ImageRequestSize;
}