# lib/api-zod/src/generated/types/modelPresets.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatModel } from "./chatModel";
import type { ImageModel } from "./imageModel";
import type { Persona } from "./persona";

export interface ModelPresets {
  chatModels: ChatModel[];
  imageModels: ImageModel[];
  personas: Persona[];
  version: string;
}