# lib/api-zod/src/generated/types/persona.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KawaiiGPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Persona {
  id: string;
  label: string;
  description: string;
  emoji: string;
}